LayoutTests/media/modern-media-controls/mute-support/mute-support-media-api-expected.txt

Testing the MuteSupport behavior by using the media API.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS mediaController.controls.muteButton.muted is false

volumechange event was dispatched
PASS media.muted is true
PASS mediaController.controls.muteButton.muted is true

volumechange event was dispatched
PASS media.muted is false
PASS mediaController.controls.muteButton.muted is false

PASS successfullyParsed is true

TEST COMPLETE
